third_party/blink/web_tests/platform/win11-arm64/external/wpt/webxr/anchors/ar_anchor_getAnchors.https-expected.txt

This is a testharness.js-based test.
[FAIL] XRFrame's trackedAnchors is empty when the feature was not requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRFrame's trackedAnchors is empty when the feature was not requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRFrame's trackedAnchors is empty when the feature was requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRFrame's trackedAnchors is empty when the feature was requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
